ajustess

+++ scriptMenu.js
@@ -97,8 +97,6 @@ document.addEventListener('DOMContentLoaded', (event) => {
                                     for (i = 0; i<100; i++){
                                         if(data[i].id == numeroPokedex){
 
-                                            console.log("Numero I : " + i)
-
                                         fetch("https://localhost:44373/api/pokemon", {
                                             method: 'DELETE',
                                             headers: {
@@ -161,85 +159,118 @@ document.addEventListener('DOMContentLoaded', (event) => {
                             var nome = data[i].name;
                             var numeroPokedex = data[i].id;
                             var sprite = data[i].frontSpriteUrl;
+                            var spriteTrazeira = data[i].backSpriteUrl
+                            var spriteShiny = data[i].frontShinySpriteUrl
+                            var spriteTrazeiraShinyt = data[i].backShinySpriteUrl
 
-
+    
                             const li = document.createElement('div');
                             li.className = "margem2 row";
-
-
+    
                             const DivImagem = document.createElement('div');
                             DivImagem.className = "col-4 text-center ";
-
+    
                             const Imagem = document.createElement('img');
-                            Imagem.className = " text-center  ";
+                            Imagem.className = "text-center";
                             Imagem.src = sprite;
-
+    
                             DivImagem.appendChild(Imagem);
-
-
+    
                             const DivId = document.createElement('div');
                             DivId.className = "col-3 text-center ";
-
+    
                             const Id = document.createElement('span');
-                            Id.className = " text-center fw-bold";
+                            Id.className = "text-center ";
                             Id.textContent = numeroPokedex;
-
+    
                             DivId.appendChild(Id);
-
-
+    
                             const DivNome = document.createElement('div');
                             DivNome.className = "col-3 text-center ";
-
+    
                             const Nome = document.createElement('span');
-                            Nome.className = " text-center fw-bold";
+                            Nome.className = "text-center";
                             Nome.textContent = nome;
-
+    
                             DivNome.appendChild(Nome);
-
-
+    
                             const DivEditar = document.createElement('div');
                             DivEditar.className = "col-1 text-center ";
-
-                            const Editar = document.createElement('a');
-                            Editar.href = "excluir.html";
-                            Editar.className = "link-offset-2 link-underline link-underline-opacity-0 text-dark";
+    
+                            const Editar = document.createElement('button');
+                            Editar.id = "btneditar";
+                            Editar.className = "btn btn-primary";
                             Editar.textContent = "Editar";
-
+    
+                            Editar.addEventListener('click', (function (idEditar) {
+                                return function () {
+                                    localStorage.setItem('idEditar', idEditar);
+                                    window.location.href = 'editar.html';
+                                    EditarCoisa();
+                                };
+                            })(numeroPokedex));
+    
                             DivEditar.appendChild(Editar);
-
-
+    
                             const DivExcluir = document.createElement('div');
                             DivExcluir.className = "col-1 text-center ";
-
+    
                             const Excluir = document.createElement('button');
                             Excluir.id = "btnexcluir";
                             Excluir.className = "btn btn-danger";
                             Excluir.textContent = "Excluir";
-
-                            /*Excluir.addEventListener('click', (function (numeroPokedexCapturado) {
-                                return function () {
-                                    fetch('https://localhost:44373/api/pokemon/' + numeroPokedexCapturado, {
-                                        method: 'DELETE',
-                                    })
-                                        .then(response => {
-                                            if (response.ok) {
-                                                console.log('Item deleted:', numeroPokedexCapturado);
-                                            } else {
-                                                console.error('Failed to delete item:', numeroPokedexCapturado);
+    
+                            Excluir.setAttribute('data-pokedex', numeroPokedex);
+    
+                            Excluir.addEventListener('click', function () {
+                                const numeroPokedex = this.getAttribute('data-pokedex'); 
+    
+                                console.log("Número da Pokédex:", numeroPokedex); 
+    
+                                fetch("https://localhost:44373/api/pokemon")
+                                    .then(response => response.json())
+                                    .then(data => {
+    
+                                        for (i = 0; i<100; i++){
+                                            if(data[i].id == numeroPokedex){
+    
+                                            fetch("https://localhost:44373/api/pokemon", {
+                                                method: 'DELETE',
+                                                headers: {
+                                                    'Content-Type': 'application/json'
+                                                },
+                                                body: JSON.stringify({id: numeroPokedex,
+                                                    name: nome,
+                                                    firstType: tipo01,
+                                                    secondType: tipo02,
+                                                    backSpriteUrl: spriteTrazeira,
+                                                    frontSpriteUrl: sprite,
+                                                    backShinySpriteUrl: spriteTrazeiraShinyt,
+                                                    frontShinySpriteUrl: spriteShiny
+                                                 }) 
+                                            })
+                                                .then(response => {
+    
+                                                        li.remove();
+                                                        alert('Pokémon excluído com sucesso!');
+        
+                                                })
+                                                .catch(error => console.error('Erro ao excluir o Pokémon:', error));
                                             }
-                                        })
-                                        .catch(error => console.error('Error:', error));
-                                };
-                            })(numeroPokedex));*/
-
+                                        }
+    
+                                    })
+                                    .catch(error => console.error('Erro ao buscar Pokémons:', error));
+                            });
+    
                             DivExcluir.appendChild(Excluir);
-
+    
                             li.appendChild(DivImagem);
                             li.appendChild(DivId);
                             li.appendChild(DivNome);
                             li.appendChild(DivEditar);
                             li.appendChild(DivExcluir);
-
+    
                             ul.appendChild(li);
                         }
                     }
@@ -313,12 +344,13 @@ document.getElementById('FormInfotmations').addEventListener('submit', (event) =
                         firstType: document.getElementById('Type01').value,
                         secondType: document.getElementById('Type02').value
                     })
-
                 })
-                    .catch(error => console.error('Erro ao atualizar Pokémon:', error));
+                .catch(error => console.error('Erro ao atualizar Pokémon:', error));
+
             } else {
                 alert('Pokémon não encontrado.');
             }
         })
         .catch(error => console.error('Erro ao buscar dados do Pokémon:', error));
 });
+
